[AWM] Change id type to int32_t

+++ include/bbque/app/working_mode.h
@@ -78,7 +78,7 @@ public:
 	/**
 	 * @see WorkingModeStatusIF
 	 */
-	inline int8_t Id() const {
+	inline int32_t Id() const {
 		return id;
 	}
